+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 Development:
+  - "attester inclusion" command finds the inclusion slot for a validator's attestation
   - "account info" with verbose option now displays participants for distributed accounts
   - fix issue where distributed account generation without a passphrase was not allowed